requirements.txt
fastapi
pydantic
sqlalchemy[asyncio]
asyncpg
httpx
python-dotenv
openrouter
typing-extensions
pytest
pytest-asyncio
alembic
databases
uvicorn
sentence-transformers
numpy
scikit-learn
python-jose[cryptography]
passlib[bcrypt]
